% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmems_catalog.R
\name{build_catalog}
\alias{build_catalog}
\title{Build full catalog table}
\usage{
build_catalog(products = NULL, progress = TRUE)
}
\arguments{
\item{products}{Optional vector of product_ids to process (NULL = all)}

\item{progress}{Show progress messages}
}
\description{
Build full catalog table
}
