% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmems_catalog_api.R
\name{build_cmems_catalog}
\alias{build_cmems_catalog}
\title{Build catalog from STAC}
\usage{
build_cmems_catalog(product_ids = NULL, progress = TRUE)
}
\arguments{
\item{product_ids}{Character vector of product IDs, or NULL for all}

\item{progress}{Show progress}
}
\description{
Build catalog from STAC
}
