% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmemsarco.R
\name{cmems_gdal_dsn}
\alias{cmems_gdal_dsn}
\title{Convert S3 HTTPS URL to GDAL Zarr DSN}
\usage{
cmems_gdal_dsn(url, array = NULL)
}
\arguments{
\item{url}{HTTPS URL to a Zarr store}

\item{array}{Optional array/variable name to access directly}
}
\value{
GDAL DSN string
}
\description{
Convert S3 HTTPS URL to GDAL Zarr DSN
}
\examples{
url <- "https://s3.waw3-1.cloudferro.com/mdl-arco-time-045/arco/PRODUCT/dataset/timeChunked.zarr"
cmems_gdal_dsn(url)
cmems_gdal_dsn(url, array = "sla")
}
