% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmemsarco.R
\name{cmems_gdal_dsn_s3}
\alias{cmems_gdal_dsn_s3}
\title{Convert URL to GDAL Zarr DSN (vsis3)}
\usage{
cmems_gdal_dsn_s3(url, array = NULL)
}
\arguments{
\item{url}{HTTPS URL to a Zarr store}

\item{array}{Optional array/variable name to access directly}
}
\value{
GDAL DSN string
}
\description{
Creates a GDAL DSN using \verb{/vsis3/} which requires \code{\link[=cmems_setup]{cmems_setup()}} to
configure AWS environment variables first.
}
\examples{
url <- "https://s3.waw3-1.cloudferro.com/mdl-arco-time-045/arco/PRODUCT/dataset/timeChunked.zarr"
cmems_setup()
cmems_gdal_dsn_s3(url)
}
