% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmemsarco.R
\name{cmems_gdalinfo_cmd}
\alias{cmems_gdalinfo_cmd}
\title{Generate gdalinfo command}
\usage{
cmems_gdalinfo_cmd(dsn)
}
\arguments{
\item{dsn}{GDAL DSN string}
}
\value{
Shell command string
}
\description{
Generate gdalinfo command
}
