% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmemsarco.R
\name{cmems_latest}
\alias{cmems_latest}
\title{Filter catalog to latest version per dataset}
\usage{
cmems_latest(catalog)
}
\arguments{
\item{catalog}{A catalog tibble from \code{\link[=cmems_catalog]{cmems_catalog()}}}
}
\value{
Filtered tibble with only the latest version of each dataset
}
\description{
Filter catalog to latest version per dataset
}
