% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmemsarco.R
\name{cmems_s3_uri}
\alias{cmems_s3_uri}
\title{Convert S3 HTTPS URL to s3:// URI}
\usage{
cmems_s3_uri(url)
}
\arguments{
\item{url}{HTTPS URL to a Zarr store}
}
\value{
S3 URI string
}
\description{
Convert S3 HTTPS URL to s3:// URI
}
