% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmemsarco.R
\name{cmems_setup}
\alias{cmems_setup}
\title{Set up GDAL environment for CMEMS S3 access}
\usage{
cmems_setup()
}
\value{
Invisible TRUE
}
\description{
Sets AWS environment variables required for \verb{/vsis3/} access to CMEMS.
Not needed if using the default \verb{/vsicurl/} DSNs.
}
\examples{
cmems_setup()
}
