% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmems_catalog_api.R
\name{get_all_products}
\alias{get_all_products}
\title{Get all products from CMEMS API
Uses the same endpoint as CopernicusMarine::cms_products_list()}
\usage{
get_all_products()
}
\description{
Get all products from CMEMS API
Uses the same endpoint as CopernicusMarine::cms_products_list()
}
