% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmems_catalog.R
\name{get_product_datasets}
\alias{get_product_datasets}
\title{Get dataset items for a product
Returns list of item IDs (dataset_version combos)}
\usage{
get_product_datasets(product_id)
}
\description{
Get dataset items for a product
Returns list of item IDs (dataset_version combos)
}
