% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmems_catalog_api.R
\name{get_product_details}
\alias{get_product_details}
\title{Get product details including layers (datasets)
This requires product-specific API call}
\usage{
get_product_details(product_id)
}
\description{
Get product details including layers (datasets)
This requires product-specific API call
}
