% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmems_catalog_api.R
\name{get_stac_datasets}
\alias{get_stac_datasets}
\title{Get datasets from STAC product file
Returns dataset items with asset URLs}
\usage{
get_stac_datasets(product_id)
}
\description{
Get datasets from STAC product file
Returns dataset items with asset URLs
}
