% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmems_catalog_api.R
\name{list_zarr_arrays}
\alias{list_zarr_arrays}
\title{List subdatasets/arrays in a Zarr
Requires vapour or similar}
\usage{
list_zarr_arrays(url)
}
\description{
List subdatasets/arrays in a Zarr
Requires vapour or similar
}
