% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmems_catalog_api.R
\name{make_arco_gdal}
\alias{make_arco_gdal}
\title{Make GDAL DSN directly}
\usage{
make_arco_gdal(
  product_id,
  dataset_id,
  version,
  chunk_type = c("time", "geo"),
  bucket_version = "045",
  array = NULL
)
}
\description{
Make GDAL DSN directly
}
