% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmems_catalog_api.R
\name{make_gdal_zarr_dsn}
\alias{make_gdal_zarr_dsn}
\title{Convert S3 HTTPS URL to GDAL Zarr DSN}
\usage{
make_gdal_zarr_dsn(url, array = NULL)
}
\arguments{
\item{url}{The https URL to zarr store}

\item{array}{Optional: specific array/variable name}
}
\description{
Convert S3 HTTPS URL to GDAL Zarr DSN
}
