% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmems_catalog.R
\name{to_gdal_zarr}
\alias{to_gdal_zarr}
\title{Generate GDAL Zarr DSN from S3 URL}
\usage{
to_gdal_zarr(s3_url, array = NULL)
}
\arguments{
\item{s3_url}{The https:// URL to the Zarr store}

\item{array}{Optional array name to access a specific variable}
}
\description{
Generate GDAL Zarr DSN from S3 URL
}
