% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmemsarco.R
\name{cmems_setup}
\alias{cmems_setup}
\title{Set up GDAL environment for CMEMS access}
\usage{
cmems_setup()
}
\value{
Invisible TRUE
}
\description{
Sets AWS environment variables required for anonymous S3 access to CMEMS.
}
\examples{
cmems_setup()
}
